Move photosensitizer CLI parser to ferroptosis-core (#205)
Issue #205 originally proposed adding --photosensitizer / --dli-h to all
"PDT-using binaries." Audit revealed those references are exhaustive
match arms, not iteration sites. The treatments arrays in sim-tme,
sim-combo, sim-combo-mech, sim-icd, sim-window, sim-invivo, sim-original,
sim-tissue-pk, and sim-tumor-pk do NOT contain Treatment::PDT. Only
sim-spatial actually iterates PDT through the photosensitizer-aware
spatial physics. Adding flags elsewhere would be dead code.

Shipping the part of #205 that is genuinely useful: extract the parser
from sim-spatial/src/main.rs into ferroptosis-core, where it lives with
the type it parses.

Changes:

- ferroptosis-core: new Photosensitizer::from_cli_spec(s) method, new
  validate_dli_h(f) free fn. Both moved verbatim from sim-spatial. Their
  17 unit tests (parse paths, value rejection, case-insensitivity,
  whitespace, Display↔parse round-trip) move with them; ferroptosis-core
  test count: 49 → 66.

- sim-spatial: replace local parse_photosensitizer_spec /
  validate_dli_h with calls to the moved versions. Test module shrinks
  from 18 → 1: only the binary-specific clap-default invariant
  (default_args_match_default_spatial_params) remains. All parser-level
  tests now live with the parser in ferroptosis-core.

- ferroptosis-core/Cargo.toml: 0.3.0 → 0.3.1 (additive patch in 0.x
  convention; no breaking changes).

- sim-tme, sim-combo, sim-combo-mech: no changes. Issue #205's premise
  about these binaries was wrong (Treatment::PDT match arms are
  exhaustiveness guards, not iteration). Documented in the issue close
  comment.

Verification:
- cargo test --workspace: ferroptosis-core 66, sim-spatial 1; all green
- pytest: 91/91
- sim-spatial 100×100 default snapshot: byte-identical (SHA-256 match)
- Display↔from_cli_spec round-trip is now a colocated test rather than
  cross-crate
